fix(wgpu): return input-dtype masks from compare and support broadcasting
WebGPU compare kernels always returned F32 masks (1.0/0.0) and rejected
mismatched shapes by falling back off the GPU entirely, diverging from
CPU and CUDA, which return the input dtype. Compare now shares the
same broadcast dispatch path as arithmetic ops (extracted into
`launch_broadcast` in wgpu/ops/helpers.rs) and its I32/U32/F32 kernels
write the input dtype instead of F32.

Backend parity tests previously only covered float dtypes and exact
shape matches, so they missed this. Compare tests now run over I32 and
U32 too, add broadcast cases, and assert mask dtype matches CPU's -
which also caught CUDA's supported-dtype list wrongly claiming U32 for
compare, cast, and binary ops it doesn't actually implement.

+++ src/runtime/wgpu/shaders/compare_i32.wgsl
@@ -1,4 +1,4 @@
-// I32 comparison operations (input I32, output F32: 1.0=true, 0.0=false)
+// I32 comparison operations (input I32, output I32: 1=true, 0=false)
 
 const WORKGROUP_SIZE: u32 = 256u;
 
@@ -8,14 +8,14 @@ struct CompareParams {
 
 @group(0) @binding(0) var<storage, read_write> compare_a: array<i32>;
 @group(0) @binding(1) var<storage, read_write> compare_b: array<i32>;
-@group(0) @binding(2) var<storage, read_write> compare_out: array<f32>;
+@group(0) @binding(2) var<storage, read_write> compare_out: array<i32>;
 @group(0) @binding(3) var<uniform> compare_params: CompareParams;
 
 @compute @workgroup_size(256)
 fn eq_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
     let idx = gid.x;
     if (idx < compare_params.numel) {
-        compare_out[idx] = select(0.0, 1.0, compare_a[idx] == compare_b[idx]);
+        compare_out[idx] = select(0, 1, compare_a[idx] == compare_b[idx]);
     }
 }
 
@@ -23,7 +23,7 @@ fn eq_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
 fn ne_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
     let idx = gid.x;
     if (idx < compare_params.numel) {
-        compare_out[idx] = select(0.0, 1.0, compare_a[idx] != compare_b[idx]);
+        compare_out[idx] = select(0, 1, compare_a[idx] != compare_b[idx]);
     }
 }
 
@@ -31,7 +31,7 @@ fn ne_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
 fn lt_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
     let idx = gid.x;
     if (idx < compare_params.numel) {
-        compare_out[idx] = select(0.0, 1.0, compare_a[idx] < compare_b[idx]);
+        compare_out[idx] = select(0, 1, compare_a[idx] < compare_b[idx]);
     }
 }
 
@@ -39,7 +39,7 @@ fn lt_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
 fn le_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
     let idx = gid.x;
     if (idx < compare_params.numel) {
-        compare_out[idx] = select(0.0, 1.0, compare_a[idx] <= compare_b[idx]);
+        compare_out[idx] = select(0, 1, compare_a[idx] <= compare_b[idx]);
     }
 }
 
@@ -47,7 +47,7 @@ fn le_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
 fn gt_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
     let idx = gid.x;
     if (idx < compare_params.numel) {
-        compare_out[idx] = select(0.0, 1.0, compare_a[idx] > compare_b[idx]);
+        compare_out[idx] = select(0, 1, compare_a[idx] > compare_b[idx]);
     }
 }
 
@@ -55,6 +55,6 @@ fn gt_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
 fn ge_i32(@builtin(global_invocation_id) gid: vec3<u32>) {
     let idx = gid.x;
     if (idx < compare_params.numel) {
-        compare_out[idx] = select(0.0, 1.0, compare_a[idx] >= compare_b[idx]);
+        compare_out[idx] = select(0, 1, compare_a[idx] >= compare_b[idx]);
     }
 }
